# backend/src/routes/emergency.ts
import { Router, Request, Response } from 'express';

// Simple logger fallback
const logger = {
  info: (...args: any[]) => console.log('[INFO]', ...args),
  error: (...args: any[]) => console.error('[ERROR]', ...args),
  debug: (...args: any[]) => console.debug('[DEBUG]', ...args),
};

export const emergencyRouter = Router();

/**
 * POST /emergency/withdraw
 * Execute emergency withdrawal
 */
emergencyRouter.post('/withdraw', async (req: Request, res: Response) => {
  try {
    const { treasury_id, recipient, amount, coin_type, reason, signatures } = req.body;

    // TODO: Implement emergency withdrawal on-chain
    logger.info(`Emergency withdrawal: treasury=${treasury_id}, recipient=${recipient}, amount=${amount}, type=${coin_type}, reason=${reason}, sigs=${signatures?.length || 0}`);
    
    return res.status(501).json({ 
      error: 'Not implemented',
      message: 'Emergency withdrawal not yet implemented' 
    });
  } catch (error: any) {
    logger.error('Error executing emergency withdrawal:', error);
    return res.status(500).json({ error: 'Failed to execute emergency withdrawal', message: error.message });
  }
});

/**
 * POST /emergency/freeze
 * Freeze treasury
 */
emergencyRouter.post('/freeze', async (req: Request, res: Response) => {
  try {
    const { treasury_id, reason } = req.body;

    // TODO: Implement treasury freeze on-chain
    logger.info(`Freeze treasury ${treasury_id}: ${reason}`);

    return res.status(501).json({ 
      error: 'Not implemented',
      message: 'Treasury freeze not yet implemented' 
    });
  } catch (error: any) {
    logger.error('Error freezing treasury:', error);
    return res.status(500).json({ error: 'Failed to freeze treasury', message: error.message });
  }
});