# backend/src/routes/policy.ts
import { Router, Request, Response } from 'express';

// Simple logger fallback
const logger = {
  info: (...args: any[]) => console.log('[INFO]', ...args),
  error: (...args: any[]) => console.error('[ERROR]', ...args),
  debug: (...args: any[]) => console.debug('[DEBUG]', ...args),
};

export const policyRouter = Router();

/**
 * POST /policy/create
 * Create a spending policy
 */
policyRouter.post('/create', async (req: Request, res: Response) => {
  try {
    const { treasury_id, policy_type, config } = req.body;

    // TODO: Implement policy creation on-chain
    logger.info(`Create policy: treasury=${treasury_id}, type=${policy_type}, config=`, config);
    
    return res.status(501).json({ 
      error: 'Not implemented',
      message: 'Policy creation not yet implemented' 
    });
  } catch (error: any) {
    logger.error('Error creating policy:', error);
    return res.status(500).json({ error: 'Failed to create policy', message: error.message });
  }
});

/**
 * GET /policy/:id
 * Get policy details
 */
policyRouter.get('/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;

    // TODO: Implement policy fetching from on-chain
    logger.debug(`Fetch policy ${id}`);

    return res.status(501).json({ 
      error: 'Not implemented',
      message: 'Policy fetching not yet implemented' 
    });
  } catch (error: any) {
    logger.error('Error fetching policy:', error);
    return res.status(500).json({ error: 'Failed to fetch policy', message: error.message });
  }
});

/**
 * DELETE /policy/:id
 * Remove a policy
 */
policyRouter.delete('/:id', async (req: Request, res: Response) => {
  try {
    const { id } = req.params;

    // TODO: Implement policy removal on-chain
    logger.info(`Remove policy ${id}`);

    return res.status(501).json({ 
      error: 'Not implemented',
      message: 'Policy removal not yet implemented' 
    });
  } catch (error: any) {
    logger.error('Error removing policy:', error);
    return res.status(500).json({ error: 'Failed to remove policy', message: error.message });
  }
});